profile and setting page updated
