+++ src/pages/Profile/Profile.types.ts
@@ -1,0 +1,10 @@
+export interface UserProfile {
+  id: string;
+  firstName: string;
+  lastName: string;
+  email: string;
+  role: string;
+  accountStatus: "active" | "suspended" | "pending";
+  createdAt: Date;
+  updatedAt: Date;
+}
